+++ packages/graz/src/actions/account.ts
@@ -1,10 +1,10 @@
 import type { OfflineDirectSigner } from "@cosmjs/proto-signing";
-import type { ChainInfo, Key, OfflineAminoSigner } from "@keplr-wallet/types";
+import type { ChainInfo, OfflineAminoSigner } from "@keplr-wallet/types";
 
 import { RECONNECT_SESSION_KEY } from "../constant";
 import { grazSessionDefaultValues, useGrazInternalStore, useGrazSessionStore } from "../store";
 import type { Maybe } from "../types/core";
-import type { WalletType } from "../types/wallet";
+import type { Key, WalletType } from "../types/wallet";
 import type { ChainId } from "../utils/multi-chain";
 import { checkWallet, getWallet, isCapsule, isWalletConnect } from "./wallet";
 
@@ -188,6 +188,8 @@ export const reconnect = async (args?: ReconnectArgs) => {
   try {
     const isWalletReady = checkWallet(_reconnectConnector || undefined);
     if (recentChains && isWalletReady && _reconnectConnector) {
+      const isWC = isWalletConnect(_reconnectConnector);
+      if (isWC) return;
       const key = await connect({
         chainId: recentChains,
         walletType: _reconnectConnector,