+++ packages/graz/src/store/index.ts
@@ -1,14 +1,14 @@
-import type { ChainInfo, Keplr, Key } from "@keplr-wallet/types";
+import type { ChainInfo, Keplr } from "@keplr-wallet/types";
 import type { CapsuleProvider } from "@leapwallet/cosmos-social-login-capsule-provider";
 import type { ISignClient, SignClientTypes } from "@walletconnect/types";
-import type { Web3ModalConfig } from "@web3modal/standalone";
+import type { WalletConnectModalConfig } from "@walletconnect/modal";
 import { create } from "zustand";
 import type { PersistOptions } from "zustand/middleware";
 import { createJSONStorage } from "zustand/middleware";
 import { persist, subscribeWithSelector } from "zustand/middleware";
 
 import type { Dictionary } from "../types/core";
-import { WalletType } from "../types/wallet";
+import { Key, WalletType } from "../types/wallet";
 
 export interface ChainConfig {
   path?: string;
@@ -20,7 +20,10 @@ export interface ChainConfig {
 }
 export interface WalletConnectStore {
   options: SignClientTypes.Options | null;
-  web3Modal?: Pick<Web3ModalConfig, "themeVariables" | "themeMode" | "privacyPolicyUrl" | "termsOfServiceUrl"> | null;
+  walletConnectModal?: Pick<
+    WalletConnectModalConfig,
+    "themeVariables" | "themeMode" | "privacyPolicyUrl" | "termsOfServiceUrl"
+  > | null;
 }
 
 export interface CapsuleConfig {
@@ -95,7 +98,7 @@ export const grazInternalDefaultValues: GrazInternalStore = {
   walletType: WalletType.KEPLR,
   walletConnect: {
     options: null,
-    web3Modal: null,
+    walletConnectModal: null,
   },
   walletDefaultOptions: null,
   _notFoundFn: () => null,